requirements.txt
python-telegram-bot
instaloader
requests